Moved @tryghost/version-notifications-data-service to @tryghost/api-version-compatibility-service
fix https://linear.app/ghost/issue/ENG-2100/version-notifications-data-service

- this package is only used by one other package
- this commit moves the lib code and tests so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
