--- ghost/version-notifications-data-service/README.md
@@ -1,3 +1,0 @@
-# Version Notifications Data Service
-
-Package-service managing data for the version notification service
